clean up performance
