+++ frontend/src/features/collections/linked-collections/index.ts
@@ -1,1 +1,3 @@
 import "./linked-collections";
+import "./linked-collections-list";
+import "./linked-collections-list-item";
